Refactor save() method in Storage.java
