+++ src/main/java/command/EventCommand.java
@@ -1,7 +1,7 @@
 package command;
 
 import converter.StringDateConverter;
-import parser.Storage;
+import storage.Storage;
 import task.Event;
 import task.TaskList;
 import ui.Ui;
