+++ src/main/java/storage/Storage.java
@@ -1,4 +1,4 @@
-package parser;
+package storage;
 
 import converter.StringDateConverter;
 import task.Deadline;
@@ -15,6 +15,7 @@ import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.Scanner;
+import java.util.StringJoiner;
 
 /**
  * Represents the file used to store task list.
@@ -82,34 +83,12 @@ public class Storage {
      */
     public void save(TaskList tasks) throws IOException {
         FileWriter fw = new FileWriter(this.filePath);
+        StringJoiner textToAdd = new StringJoiner(System.lineSeparator());
         for (Task task : tasks.getTasks()) {
-            String description;
-            String textToAdd;
-            Date date = null;
-            textToAdd = "";
-            description = task.getDescription();
-            if (task instanceof Deadline) {
-                textToAdd += "D";
-                Deadline deadline = (Deadline) task;
-                date = deadline.getBy();
-            } else if (task instanceof Event) {
-                textToAdd += "E";
-                Event event = (Event) task;
-                date = event.getAt();
-            } else {
-                textToAdd += "T";
-            }
-            if (task.getDoneIcon().equals("\u2713")) {
-                textToAdd += " | Done";
-            } else {
-                textToAdd += " | Not done";
-            }
-            textToAdd += " | " + description.trim();
-            if (date != null) {
-                textToAdd += " | " + date;
-            }
-            fw.write(textToAdd + "\r\n");
+            textToAdd.add(task.toSaveFormat());
         }
+        fw.write(textToAdd.toString());
+        fw.write(System.lineSeparator());
 
         fw.close();
     }
